--- a/base58.go
+++ b/base58.go
@@ -1,5 +1,5 @@
 // Copyright (c) 2013-2015 The btcsuite developers
-// Copyright (c) 2015 The Decred developers
+// Copyright (c) 2015-2019 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
@@ -15,35 +15,59 @@ var bigRadix = big.NewInt(58)
 var bigZero = big.NewInt(0)
 
 // Decode decodes a modified base58 string to a byte slice.
-func Decode(b string) []byte {
-	answer := big.NewInt(0)
-	j := big.NewInt(1)
+func Decode(input string) []byte {
+	if len(input) == 0 {
+		return []byte("")
+	}
 
-	scratch := new(big.Int)
-	for i := len(b) - 1; i >= 0; i-- {
-		tmp := b58[b[i]]
-		if tmp == 255 {
+	// The max possible output size is when a base58 encoding consists of
+	// nothing but the alphabet character at index 0 which would result in the
+	// same number of bytes as the number of input chars.
+	output := make([]byte, len(input))
+
+	// Encode to base256 in reverse order to avoid extra calculations to
+	// determine the final output size in favor of just keeping track while
+	// iterating.
+	var index int
+	for _, r := range input {
+		// Invalid base58 character.
+		val := uint32(b58[r])
+		if val == 255 {
 			return []byte("")
 		}
-		scratch.SetInt64(int64(tmp))
-		scratch.Mul(j, scratch)
-		answer.Add(answer, scratch)
-		j.Mul(j, bigRadix)
-	}
 
-	tmpval := answer.Bytes()
-
-	var numZeros int
-	for numZeros = 0; numZeros < len(b); numZeros++ {
-		if b[numZeros] != alphabetIdx0 {
-			break
+		// Multiply each byte in the output by 58 and encode to base256 while
+		// propagating the carry.
+		for i, b := range output[:index] {
+			val += uint32(b) * 58
+			output[i] = byte(val)
+			val >>= 8
+		}
+		for ; val > 0; val >>= 8 {
+			output[index] = byte(val)
+			index++
 		}
 	}
-	flen := numZeros + len(tmpval)
-	val := make([]byte, flen)
-	copy(val[numZeros:], tmpval)
 
-	return val
+	// Account for the leading zeros in the input.  They are appended since the
+	// encoding is happening in reverse order.
+	for _, r := range input {
+		if r != alphabetIdx0 {
+			break
+		}
+
+		output[index] = 0
+		index++
+	}
+
+	// Truncate the output buffer to the actual number of decoded bytes and
+	// reverse it since it was calculated in reverse order.
+	output = output[:index:index]
+	for i := 0; i < index/2; i++ {
+		output[i], output[index-1-i] = output[index-1-i], output[i]
+	}
+
+	return output
 }
 
 // Encode encodes a byte slice to a modified base58 string.