--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module github.com/decred/base58
 
+go 1.11
+
 require github.com/decred/dcrd/crypto/blake256 v1.0.0