Profiling has shown the nearly 6.5% of the total allocations in dcrd are the result of `CheckDecode`, most of which are via `Decode`.

Consequently, this optimizes both to reduce memory allocations and also speed them up quite significantly for the typical usage patterns.

It consists of several commits to help ease the review.

The first commit consolidate tests into the main package to make it more consistent with the rest of the code base and to make it easier for forks since they don't have to change the import paths as much.

The second commit adds a benchmark for `CheckDecode` and updates the benchmarks for `Encode` and `Decode` to be more representative of typical usage.

The third commit updates `CheckDecode` to make use of the new zero alloc `blake256.Sum256` function to reduce the memory allocations needed to perform the decode and improves the readability of the function a bit while updating.

The final commit updates rewrites the `Decode` function to reduce the memory allocations needed to perform the decode, significantly speed it up for the typical usage patterns, and also improves the readability of the function while updating.

The following is a before and after comparison of the cumulative difference for a typical decode including the results of `CheckDecode` since it uses `Decode`.

```
benchmark               old ns/op    new ns/op    delta
---------------------------------------------------------
BenchmarkBase58Decode   4077         1036         -74.59%
BenchmarkCheckDecode    4317         1397         -67.64%

benchmark               old allocs   new allocs   delta
---------------------------------------------------------
BenchmarkBase58Decode   8            1            -87.50%
BenchmarkCheckDecode    12           2            -83.33%

benchmark               old bytes    new bytes    delta
---------------------------------------------------------
BenchmarkBase58Decode   232          48           -79.31%
BenchmarkCheckDecode    472          80           -83.05%
```